shared intersection observer
