+++ components/useInView.ts
@@ -1,0 +1,29 @@
+import { useCallback, useEffect, useRef, useState } from "react";
+
+export default function useInView() {
+  const [inView, setInView] = useState(false);
+  const elementRef = useRef(null);
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          setInView(entry.isIntersecting);
+        });
+      },
+      {
+        threshold: 1.0,
+      },
+    );
+
+    if (elementRef.current) {
+      observer.observe(elementRef.current);
+    }
+
+    return () => {
+      observer.disconnect();
+    };
+  }, [setInView]);
+
+  return { inView, elementRef };
+}
